refactor(back): 금액 필드 타입을 long으로 통일 - Subscription 비용 필드(totalCost/userShareCost/monthlyShareCost) long 전환 - SubscriptionEvaluation annualWaste 및 DashboardResponse 집계 금액 long 전환 - 대시보드 집계 스트림(mapToLong) 및 테스트 숫자 타입 정합성 수정

+++ backend/src/main/java/com/back/domain/dashboard/dto/DashboardResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public record DashboardResponse(
         long totalMonthlyCost,          // 해당 월 "평가된 구독" 기준 월 환산 사용자 부담금 합계
-        int totalAnnualWasteEstimate,  // 해당 월 평가된 구독 기준 연간 낭비 추정 합계
+        long totalAnnualWasteEstimate,  // 해당 월 평가된 구독 기준 연간 낭비 추정 합계
         List<SubscriptionSummary> subscriptions
 ) {
     public record SubscriptionSummary(
@@ -14,6 +14,6 @@ public record DashboardResponse(
             String name,
             double efficiencyRate,
             EvaluationStatus status,
-            int annualWaste
+            long annualWaste
     ) {}
 }
